Rename "FormatField" to "Field"
