+++ src/globus_cli/termio/output_formatter.py
@@ -8,6 +8,7 @@ from globus_cli.utils import CLIStubResponse
 
 from .awscli_text import unix_formatted_print
 from .context import get_jmespath_expression, outformat_is_json, outformat_is_unix
+from .field import Field
 
 FORMAT_SILENT = "silent"
 FORMAT_JSON = "json"
@@ -16,66 +17,6 @@ FORMAT_TEXT_RECORD = "text_record"
 FORMAT_TEXT_RECORD_LIST = "text_record_list"
 FORMAT_TEXT_RAW = "text_raw"
 FORMAT_TEXT_CUSTOM = "text_custom"
-
-
-class FormatField:
-    """A field which will be shown in record or table output.
-    When fields are provided as tuples, they are converted into this.
-
-    :param name: the displayed name for the record field or the column
-        name for table output
-    :param key: a str for indexing into print data or a callable which
-        produces a string given the print data
-    :param wrap_enabled: in record output, is this field allowed to wrap
-    """
-
-    def __init__(self, name, key, wrap_enabled=False):
-        self.name = name
-        self.keyfunc = _key_to_keyfunc(key)
-        self.wrap_enabled = wrap_enabled
-
-    @classmethod
-    def coerce(cls, rawfield):
-        """given a (FormatField|tuple), convert to a FormatField"""
-        if isinstance(rawfield, cls):
-            return rawfield
-        elif isinstance(rawfield, tuple):
-            if len(rawfield) == 2:
-                return cls(rawfield[0], rawfield[1])
-            raise ValueError("cannot coerce tuple of bad length")
-        raise TypeError(
-            "FormatField.coerce must be given a field or tuple, "
-            "got {}".format(type(rawfield))
-        )
-
-    def __call__(self, data):
-        """extract the field's value from the print data"""
-        return self.keyfunc(data)
-
-
-def _key_to_keyfunc(k):
-    """
-    We allow for 'keys' which are functions that map columns onto value
-    types -- they may do formatting or inspect multiple values on the
-    object. In order to support this, wrap string keys in a simple function
-    that does the natural lookup operation, but return any functions we
-    receive as they are.
-    """
-    # if the key is a string, then the "keyfunc" is just a basic lookup
-    # operation -- return that
-    if isinstance(k, str):
-        subkeys = k.split(".")
-
-        def lookup(x):
-            current = x
-            for sub in subkeys:
-                current = x[sub]
-            return current
-
-        return lookup
-    # otherwise, the key must be a function which is executed on the item
-    # to produce a value -- return it verbatim
-    return k
 
 
 def _jmespath_preprocess(res):
@@ -232,7 +173,7 @@ def formatted_print(
     must take ``response_data`` and produce another dict or dict-like object
     (json/unix output only)
 
-    ``fields`` is an iterable of fields. They may be expressed as FormatField
+    ``fields`` is an iterable of fields. They may be expressed as Field
     objects, (fieldname, key_string) tuples, or (fieldname, key_func) tuples.
 
     ``response_key`` is a key into the data to print. When used with table
@@ -306,9 +247,9 @@ def formatted_print(
         if text_epilog is not None:
             click.echo(text_epilog)
 
-    # ensure fields are FormatField instances
+    # ensure fields are Field instances
     if fields:
-        fields = [FormatField.coerce(f) for f in fields]
+        fields = [Field.coerce(f) for f in fields]
 
     if isinstance(text_format, str):
         text_format = text_format
